Acréscimo validação nos endpoints controller Livros

+++ Biblioteca.Api/Biblioteca.Api/Biblioteca.Api/Controllers/LivrosController.cs
@@ -1,4 +1,20 @@
-﻿using Biblioteca.Api.Models.Entities;
+﻿
+/*
+    Verbos HTTP:
+    GET = Requisição de dados ao servidor
+    POST = Inclusão de dados 
+    PUT = atualização de dados
+    DELETE = exclusão
+
+    Status:
+    200 OK 
+    204 No Content, OK sem retornar dados novamente (no PUT, DELETE)
+    400 Bad Request, validação, formatação
+    404 Not Found, ex. ID não encontrado na base
+    500 Internal Server Error, erro interno, erro de programação
+*/
+
+using Biblioteca.Api.Models.Entities;
 using Biblioteca.Api.Service;
 using Microsoft.AspNetCore.Mvc;
 
@@ -25,7 +41,14 @@ namespace Biblioteca.Api.Controllers
         [HttpGet("{id}")]
         public async Task<ActionResult<Livro>> GetLivro(int id)
         {
-            return await _service.BuscarPorIdAsync(id);
+            if (id <= 0)
+                return BadRequest("O código do livro deve ser maior que 0.");
+
+            var livro = await _service.BuscarPorIdAsync(id);
+            if (livro == null)
+                return NotFound();
+
+            return Ok(livro);
         }
 
         [HttpPost]
@@ -50,6 +73,9 @@ namespace Biblioteca.Api.Controllers
         [HttpDelete("{id}")]
         public async Task<IActionResult> DeleteLivro(int id)
         {
+            if (id <= 0)
+                return BadRequest("O ID deve ser maior do que 0.");
+
             var removido = await _service.RemoverLivroAsync(id);
             if (!removido)
                 return NotFound();
